+++ backend/baguni-api/src/main/java/baguni/api/application/ranking/dto/RankingApiMapper.java
@@ -5,7 +5,7 @@ import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.ReportingPolicy;
 
-import baguni.api.service.link.dto.LinkInfo;
+import baguni.entity.model.link.dto.LinkInfo;
 import baguni.common.dto.UrlWithCount;
 
 @Mapper(
